requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
anthropic>=0.84.0
weasyprint>=68.0
Pillow>=11.0.0
Jinja2>=3.1.4
numpy>=2.0.0
